* Fixed timer overflow
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@86 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gavl/gavl/timer.c
@@ -10,7 +10,7 @@ static void get_time(gavl_time_t * ret)
   {
   struct timeval time;
   gettimeofday(&time, NULL);
-  *ret = time.tv_sec*1000000 + time.tv_usec;
+  *ret = (int64_t)(time.tv_sec)*1000000LL + time.tv_usec;
   }
 
 struct gavl_timer_s
